w3c-reports/12092014/mac/runned-tests/2dcontext/compositing/2d.composite.operation.darker-expected.txt

2d.composite.operation.darker
Actual output:

FAIL Canvas test: 2d.composite.operation.darker assert_equals: ctx.globalCompositeOperation === 'xor' (got darker[string], expected xor[string]) expected "xor" but got "darker"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
_assertSame@http://localhost:8800/common/canvas-tests.js:17:18
http://localhost:8800/2dcontext/compositing/2d.composite.operation.darker.html:24:12
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/common/canvas-tests.js:62:19
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/common/canvas-tests.js:59:15)
